Determine descriptor similarity.
This change makes it possible to compute the similarity between all
router descriptors in a given router descriptor file.  That results in
approximately n^2/2 similarities.  The similarities can be used to find
groups of relays in the Tor network.

At the moment, the similarity is only a bunch of numbers.  Ultimately,
it would be great if we could define a mathematical metric because it
would allow us to process files with less comparisons.