Implement vantage point tree search.
Vantage point trees enable searching for nearest neighbours in O(log n).
They require, however, a distance metric in the mathematical sense.  For
now, we use the Levenshtein (a.k.a. edit) distance.  While it is hard to
incorporate domain-specific knowledge, the Levenshtein distance happily
takes preprocessed router descriptors as input, which is a good start.