Add module for uptime visualisation.
The approach is very similar to:
<https://trac.torproject.org/projects/tor/ticket/12813>
The idea is to take a bunch of consensuses, and plot the uptime pattern
of all relays in the consensuses.  The x-axis is the relay index, i.e.,
one column for every relay.  The y-axis is one consensus, and is either
white (relay offline) or black (relay online).  Visually similar blocks
are marked as red, based on a distance function.  The code can be used
as follows:

sybilhunter -data /path/to/consensuses/ -uptime

What's still missing, however, is a way to easily figure out which
relays are in a "relay block" in the image.