Use two time series and add smoother.
If we combine relays that appear and disappear into a single churn
value, the numbers could cancel each other out and we might miss
significant events.  As a result, we split it into two time series: one
for relays that appear, and one for relays that disappear.

This commit also adds code for a moving average that smoothes the time
series and facilitates detecting trend changes.