Use clustering to sort uptime visualisation.
Previously, we would sort the columns of our uptime visualisation based
on their median.  This didn't work so well because a similar median does
not imply a similar uptime pattern.  This commit changes our sorting to
single-linkage clustering with the Pearson correlation coefficient as
distance function.  As a result, columns that are highly correlated are
grouped together.  This achieves significantly better results.

The approach is identical to Section 2 of the following paper:
<http://nms.csail.mit.edu/papers/clustering-imw2002.pdf>